Adding post_build
