+++ pyproject.toml
@@ -1,3 +1,3 @@
 [build-system]
-requires = ["setuptools >= 40.6.0", "wheel", "lyncs_setuptools"]
+requires = ["setuptools >= 40.6.0", "wheel", "lyncs_setuptools", "cppyy", "black"]
 build-backend = "setuptools.build_meta"
